payment_service/pyproject.toml
[project]
name = "payment-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiokafka>=0.12.0",
    "aiosqlite>=0.21.0",
    "alembic>=1.16.4",
    "asyncpg>=0.30.0",
    "fastapi>=0.116.1",
    "grpcio>=1.74.0",
    "grpcio-tools>=1.74.0",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "python-dotenv>=1.1.1",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.41",
    "starlette-exporter>=0.23.0",
    "stripe>=12.3.0",
    "uvicorn>=0.35.0",
]
